Reorganize election as a feature
