+++ fe1-web/features/evoting/network/messages/EndElection.ts
@@ -1,10 +1,8 @@
-import {
-  Hash, Timestamp,
-} from 'model/objects';
+import { Hash, Timestamp } from 'model/objects';
 import { ProtocolError } from 'model/network/ProtocolError';
 import { validateDataObject } from 'model/network/validation';
-import { ActionType, MessageData, ObjectType } from '../MessageData';
-import { checkTimestampStaleness } from '../Checker';
+import { ActionType, MessageData, ObjectType } from 'model/network/method/message/data/MessageData';
+import { checkTimestampStaleness } from 'model/network/method/message/data/Checker';
 
 /** Data sent to end an Election event */
 export class EndElection implements MessageData {
