+++ fe1-web/features/evoting/objects/index.ts
@@ -1,0 +1,1 @@
+export * from './Election';
