+++ fe1-web/ingestion/handlers/index.ts
@@ -2,9 +2,9 @@ import { MessageRegistry } from 'model/network/method/message/data/MessageRegist
 import * as ChirpHandler from 'features/social/network/ChirpHandler';
 import * as ReactionHandler from 'features/social/network/ReactionHandler';
 import * as MeetingHandler from 'features/meeting/network/MeetingHandler';
+import * as ElectionHandler from 'features/evoting/network/ElectionHandler';
 import * as LaoHandler from './LaoHandler';
 import * as RollCallHandler from './RollCallHandler';
-import * as ElectionHandler from './ElectionHandler';
 import * as WitnessHandler from './WitnessHandler';
 
 type ConfigurableHandler = {
