+++ fe1-web/model/network/method/message/data/index.ts
@@ -3,5 +3,4 @@ export * from './Checker';
 export * from './witness';
 export * from './lao';
 export * from './rollCall';
-export * from './election';
 export * from './MessageRegistry';
