+++ fe1-web/model/objects/index.ts
@@ -5,7 +5,6 @@ export const Wallet = w;
 export * from './Base64Url';
 export * from './Channel';
 export * from './ConnectToLao';
-export * from './Election';
 export * from './Hash';
 export * from './KeyPair';
 export * from './Lao';
