+++ fe1-web/network/MessageApi.ts
@@ -2,14 +2,11 @@ import {
   EventTags, Hash, Lao, PublicKey, Timestamp,
 } from 'model/objects';
 import {
-  CastVote,
   CloseRollCall,
   CreateLao,
   CreateRollCall,
-  EndElection,
   OpenRollCall,
   ReopenRollCall,
-  SetupElection,
   StateLao,
   UpdateLao,
   WitnessMessage,
@@ -20,7 +17,6 @@ import {
 import {
   OpenedLaoStore, KeyPairStore,
 } from 'store';
-import { Question, Vote } from 'model/objects/Election';
 import { publish } from './JsonRpcApi';
 
 /**
@@ -194,70 +190,4 @@ export function requestCloseRollCall(
   });
 
   return publish(channelFromIds(lao.id), message);
-}
-
-/** Sends a server query asking for creation of an Election with a given name (String),
- *  an array of questions, a version (String), the current lao (String), the id, and the  creation,
- *  start and end are also specified as a timestamp */
-export function requestCreateElection(
-  name: string,
-  version: string,
-  start: Timestamp,
-  end: Timestamp,
-  questions: Question[],
-  time: Timestamp,
-): Promise<void> {
-  const currentLao: Lao = OpenedLaoStore.get();
-
-  const message = new SetupElection({
-    lao: currentLao.id,
-    id: Hash.fromStringArray(
-      EventTags.ELECTION, currentLao.id.toString(), time.toString(), name,
-    ),
-    name: name,
-    version: version,
-    created_at: time,
-    start_time: adaptStartTime(time, start),
-    end_time: end,
-    questions: questions,
-  });
-
-  const laoCh = channelFromIds(currentLao.id);
-  return publish(laoCh, message);
-}
-
-/** Sends a server query which creates a Vote in an ongoing election */
-export function castVote(
-  election_id: Hash,
-  votes: Vote[],
-): Promise<void> {
-  const time: Timestamp = Timestamp.EpochNow();
-  const currentLao: Lao = OpenedLaoStore.get();
-  const message = new CastVote({
-    lao: currentLao.id,
-    election: election_id,
-    created_at: time,
-    votes: votes,
-  });
-
-  const elecCh = channelFromIds(currentLao.id, election_id);
-  return publish(elecCh, message);
-}
-
-/** Sends a server query which creates a Vote in an ongoing election */
-export function terminateElection(
-  electionId: Hash,
-  registeredVotes: Hash,
-): Promise<void> {
-  const time: Timestamp = Timestamp.EpochNow();
-  const currentLao: Lao = OpenedLaoStore.get();
-  const message = new EndElection({
-    lao: currentLao.id,
-    election: electionId,
-    created_at: time,
-    registered_votes: registeredVotes,
-  });
-
-  const elecCh = channelFromIds(currentLao.id, electionId);
-  return publish(elecCh, message);
 }
